+++ tests/__init__.py
@@ -1,0 +1,1 @@
+"""Local test support package used by documentation examples."""